+++ main_test.go
@@ -136,7 +136,7 @@ func TestMain(m *testing.M) {
 	jail.JailsRootDir = "./jails_test"
 
 	now := time.Now()
-	for i := 0; i < len(testTokens); i++ {
+	for i := int32(0); i < int32(len(testTokens)); i++ {
 		testTokens[i], err = GenerateSessionToken()
 		if err != nil {
 			log.Fatalf("Failed to generate session token: %v", err)
